+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.2 (unreleased)
 
 - Added `p_iter` and `q_iter`
+- Added `p_row` and `q_col`
 
 ## 0.2.1 (2021-11-15)